move all Polychat code to Mmcc.Bot.Polychat
