--- src/Mmcc.Bot.Core/Extensions/OnlineServerExtensions.cs
@@ -1,9 +1,0 @@
-﻿namespace Mmcc.Bot.Core.Extensions
-{
-    public static class OnlineServerExtensions
-    {
-        public static OnlineServerInformation ExtractServerInformation(this OnlineServer onlineServer) =>
-            new(onlineServer.ServerId, onlineServer.ServerName, onlineServer.ServerAddress,
-                onlineServer.MaxPlayers, onlineServer.PlayersOnline, onlineServer.OnlinePlayerNames);
-    }
-}
