+++ src/Mmcc.Bot.Core/Mmcc.Bot.Core.csproj
@@ -15,7 +15,6 @@
     <ItemGroup>
       <ProjectReference Include="..\Mmcc.Bot.Caching\Mmcc.Bot.Caching.csproj" />
       <ProjectReference Include="..\Mmcc.Bot.Database\Mmcc.Bot.Database.csproj" />
-      <ProjectReference Include="..\Mmcc.Bot.Protos\Mmcc.Bot.Protos.csproj" />
     </ItemGroup>
 
 </Project>
