+++ src/Mmcc.Bot.Infrastructure/Commands/MemberApplications/ApproveAutomatically.cs
@@ -9,7 +9,7 @@ using Mmcc.Bot.Core.Errors;
 using Mmcc.Bot.Database;
 using Mmcc.Bot.Database.Entities;
 using Mmcc.Bot.Infrastructure.Services;
-using Mmcc.Bot.Protos;
+using Mmcc.Bot.Polychat;
 using Remora.Discord.API.Abstractions.Rest;
 using Remora.Discord.Core;
 using Remora.Results;
