+++ src/Mmcc.Bot.Infrastructure/Commands/ModerationActions/Ban.cs
@@ -11,7 +11,7 @@ using Mmcc.Bot.Core.Statics;
 using Mmcc.Bot.Database;
 using Mmcc.Bot.Database.Entities;
 using Mmcc.Bot.Infrastructure.Services;
-using Mmcc.Bot.Protos;
+using Mmcc.Bot.Polychat;
 using Remora.Discord.API.Abstractions.Rest;
 using Remora.Discord.API.Objects;
 using Remora.Discord.Core;
