+++ src/Mmcc.Bot.Infrastructure/Mmcc.Bot.Infrastructure.csproj
@@ -19,7 +19,7 @@
     <ItemGroup>
       <ProjectReference Include="..\Mmcc.Bot.Core\Mmcc.Bot.Core.csproj" />
       <ProjectReference Include="..\Mmcc.Bot.Database\Mmcc.Bot.Database.csproj" />
-      <ProjectReference Include="..\Mmcc.Bot.Protos\Mmcc.Bot.Protos.csproj" />
+      <ProjectReference Include="..\Mmcc.Bot.Polychat\Mmcc.Bot.Polychat.csproj" />
     </ItemGroup>
 
     <ItemGroup>
