+++ src/Mmcc.Bot.Polychat/Abstractions/PolychatChatMessageString.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Text.RegularExpressions;
 
-namespace Mmcc.Bot.Core
+namespace Mmcc.Bot.Polychat.Abstractions
 {
     /// <summary>
     /// Represents a polychat chat message string.
