+++ src/Mmcc.Bot.Polychat/IncomingMessageHandlers/HandleServerInfo.cs
@@ -1,10 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
-using Mmcc.Bot.Core;
-using Mmcc.Bot.Infrastructure.Services;
-using Mmcc.Bot.Protos;
+using Mmcc.Bot.Polychat.Abstractions;
+using Mmcc.Bot.Polychat.Models;
 
-namespace Mmcc.Bot.Infrastructure.Commands.Polychat.IncomingMessageHandlers
+namespace Mmcc.Bot.Polychat.IncomingMessageHandlers
 {
     /// <summary>
     /// Handles an incoming <see cref="ServerInfo"/> message.
