+++ src/Mmcc.Bot/CommandGroups/Core/MmccInfoCommands.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.ComponentModel;
 using System.Threading.Tasks;
 using Mmcc.Bot.Caching;
-using Mmcc.Bot.Caching.Entities;
 using Mmcc.Bot.Core.Extensions.Caching;
 using Mmcc.Bot.Core.Statics;
 using Mmcc.Bot.RemoraAbstractions;
