+++ src/Mmcc.Bot/CommandGroups/Minecraft/MinecraftServersCommands.cs
@@ -7,9 +7,10 @@ using System.Threading.Tasks;
 using MediatR;
 using Mmcc.Bot.Core.Models;
 using Mmcc.Bot.Core.Statics;
-using Mmcc.Bot.Infrastructure.Commands.Polychat.MessageSenders;
 using Mmcc.Bot.Infrastructure.Conditions.Attributes;
 using Mmcc.Bot.Infrastructure.Services;
+using Mmcc.Bot.Polychat;
+using Mmcc.Bot.Polychat.MessageSenders;
 using Mmcc.Bot.RemoraAbstractions;
 using Remora.Commands.Attributes;
 using Remora.Commands.Groups;
