+++ src/Mmcc.Bot/Program.cs
@@ -26,10 +26,11 @@ using Mmcc.Bot.Infrastructure.HostedServices;
 using Mmcc.Bot.Infrastructure.Parsers;
 using Mmcc.Bot.Infrastructure.Queries.Core;
 using Mmcc.Bot.Infrastructure.Services;
+using Mmcc.Bot.Polychat;
+using Mmcc.Bot.Polychat.Models.Settings;
 using Mmcc.Bot.Responders.Guilds;
 using Mmcc.Bot.Responders.Messages;
 using Mmcc.Bot.Responders.Users;
-using Mmcc.Bot.Protos;
 using Mmcc.Bot.RemoraAbstractions;
 using Mmcc.Bot.Responders.Buttons;
 using Remora.Commands.Extensions;
